# server/routes/apiDocumentation.yaml
openapi: 3.0.1
info:
    title: Ecommerce API (Codecademy)
    version: 1.0.0
    description: An ecommerce API
paths:
    /users/{id}:
        summary: Get user data
        description: Used to retrieve user data from DB
        get:
            summary: Get user by user ID
            description: Returns user basic info
            tags:
                - Users
            parameters:
                - name: id
                  in: path
                  required: true
                  description: The ID of the user to retrieve
                  schema:
                    type: integer
            responses:
                '200':
                    description: Info on user
                    content:
                        application/json:
                            example:
                                id: 1
                                email: "johndoe@gmail.com"
                                firstname: "John"
                                lastname: "Doe"
        put:
            summary: Update a user by user id
            description: Update a users info and return the basic user info
            tags:
                - Users
            parameters:
                - name: id
                  in: path
                  required: true
                  description: The ID of the user to update
                  schema:
                    type: integer
            requestBody:
                description: User data to update
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                firstname:
                                    type: string
                                lastname:
                                    type: string
                        example:
                            email: "johndoe1@gmail.com"
            responses:
                '200':
                    description: Update a user
                    content:
                        application/json:
                            example:
                                email: "johndoe1@gmail.com"
                                firstname: "John"
                                lastname: "Doe"

    /orders:
        summary: Getall of the order data.
        description: This path is used to retrieve order data from the orders DB
        get:
            summary: Get all orders
            description: Returns a list of all orders.
            tags: 
                - Orders
            responses:
                '200':
                    description: A list of orders.
                    content:
                        application/json:
                            example:
                                - "id": 68
                                  "created": "2026-06-26T02:35:47.707Z"
                                  "modified": "2026-06-26T02:35:47.707Z"
                                  "total": "1299.97"
                                  "status": "PENDING"
                                  "userid": 1
                                - "id": 75
                                  "created": "2026-06-27T23:35:18.757Z"
                                  "modified": "2026-06-27T23:35:20.763Z"
                                  "total": "499.99"
                                  "status": "COMPLETE"
                                  "userid": 3
    /orders/user/{userId}: 
        summary: Get orders by user ID
        description: This path is used to retrieve all orders for a particular user
        get:
            summary: Get user order
            description: Retrieves orders for user
            tags: 
                - Orders
            parameters:
                - name: userId
                  in: path
                  required: true
                  description: The ID of the user
                  schema:
                    type: integer
            responses:
                '200':
                    description: Successfully retrieved orders for user
                    content:
                        application/json:
                            example:
                                "id": 75
                                "created": "2026-06-27T23:35:18.757Z"
                                "modified": "2026-06-27T23:35:20.763Z"
                                "total": "499.99"
                                "status": "COMPLETE"
                                "userid": 3
                                "items":
                                    "productid": 5
                                    "quantity": 2
                                    "price": 499.99
                                    "name": "MacBook Neo"
                                    "description": "Entry Level computer"
                                
    /orders/{id}:
        summary: Retrieve an order
        description: This path is used to retrieve an order by the ID from the orders DB
        get:
            summary: Get order
            description: Retrieve order by order ID
            tags: 
                - Orders
            parameters:
              - name: id
                in: path
                required: true
                schema:
                    type: string
                description: The ID of the order.
            responses:
                '200':
                    description: Order successfully retrieved.
                    content:
                        application/json:
                            example:
                                "id": 68
                                "created": "2026-06-26T02:35:47.707Z"
                                "modified": "2026-06-26T02:35:47.707Z"
                                "total": "1299.97"
                                "status": "PENDING"
                                "userid": 1
                '404':
                    description: Order not found.
    /products:
        summary: Retrieve products of ecommerce store
        description: This path is used to retrieve all products within the ecommerce store
        get:
            summary: Get all products
            description: Retrieve all products
            tags:
                - Products
            responses:
                '200':
                    description: Products successfully retrieved.
                    content:
                        application/json: 
                          example:
                            - id: 1
                              name: "PlayStation 2"
                              price: "299.99"
                              description: "Unlocking new possibilities"
                            - id: 2
                              name: "PlayStation 3"
                              price: "399.99"
                              description: "Enter a new world of gaming"
                            - id: 3
                              name: "MacBook Pro"
                              price: "1999.99"
                              description: "For the creative professional"
                '404':
                    description: Not found
    /products/{id}:
        summary: Retrieve a product based on product ID
        description: Used to retrieve a particular product using the product's ID
        get:
            summary: Get product by product ID
            description: Retrieve product using product ID
            tags:
                - Products
            parameters:
              - name: id
                in: path
                description: Product ID
                required: true
                schema:
                    type: string
            responses:
                '200':
                    description: Product successfully retrieved.
                    content:
                        application/json:
                            example:
                              id: 3
                              name: "MacBook Pro"
                              price: "1999.99"
                              description: "For the creative professional"
                '404':
                    description: Product not found
    /carts/{id}:
        summary: Get a user's cart
        description: Retrieve user's cart using cart ID
        get:
            summary: Get cart by ID
            description: Retrieve cart using cart ID
            tags:
                - Carts
            parameters:
                - name: id
                  in: path
                  description: Cart ID
                  required: true
                  schema:
                    type: string
            responses:
                '200':
                    description: Cart successfully retrieved.
                    content:
                        application/json:
                            example:
                                cartid: 7
                                created: "2026-05-22T15:39:52.242Z"
                                modified: "2026-05-25T02:07:55.257Z"
                                userid: 2
                                items:
                                    qty: 1
                                    productid: 10
                                    name: "2004 Ford Taurus"
                                    price: "12999.99"
                                    description: "The Ultimate Family Cruiser"
                '404':
                    description: Cart not found.
    /carts/user/{userId}:
        summary: User's cart by user ID
        description: Retrieve user's cart using user ID
        get:
            summary: Get cart by user ID
            description: Retrieve cart using user ID
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  description: Cart ID
                  required: true
                  schema:
                    type: string
            responses:
                '200':
                    description: Cart successfully retrieved.
                    content:
                        application/json:
                            example:
                                cartid: 8
                                userid: 3
                                items:
                                    qty: 11
                                    productid: 1
                                    name: "PlayStation 2"
                                    price: "199.99"
                                    description: "Enter a new world of gaming"
                '404':
                    description: No cart for this user.
        post:
            summary: Create cart for user
            description: Create a new cart associated with the specified user ID
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  description: The ID of the user to create a cart for
                  required: true
                  schema:
                    type: string
            responses:
                '201':
                    description: Cart successfully created.
                    content:
                        application/json:
                            example:
                                cartid: 13
                                userid: 15
                                created: "2026-07-17T01:41:29.001Z"
                                modified: "2026-07-17T01:41:29.001Z"
                '404':
                    description: No cart in database
    /carts/user/{userId}/items:
        summary: Add items for user's cart
        description: Add items to user's cart
        post:
            summary: Add items to cart
            description: Add items to user's cart
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  description: The ID of the user to add items to cart
                  required: true
                  schema:
                    type: string
            responses:
                '201':
                    description: Item(s) successfully added to cart.
                    content:
                        application/json:
                            example:
                                cartid: 4
                                productid: 9
                                qty: 1
                '400':
                    description: No items to add
        delete:
            summary: Empty user cart
            description: Deletes all items in the cart
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  required: true
                  description: The ID of the user whose cart should be emptied.
                  schema:
                    type: string
            responses:
                '204':
                    description: Cart successfully cleared.
                '400':
                    description: Bad request. Invalid user ID format.
                '404':
                    description: Cart or user not found.
    /carts/user/{userId}/items/{cartItemId}:
        summary: Update select item by item ID
        description: Update an item by the item's ID
        patch:
            summary: Update cart item
            description: Updates a select item by its unique item ID
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  required: true
                  description: The ID of the user.
                  schema:
                    type: string
                - name: cartItemId
                  in: path
                  required: true
                  description: The unique ID of the item in the cart to update.
                  schema:
                    type: string
            requestBody:
                description: The cart item data to update.
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                qty:
                                    type: integer
                                    description: The new quantity for the item.
                        example:
                            qty: 5
            responses:
                '200':
                    description: Cart item successfully updated.
                    content:
                        application/json:
                            example:
                                id: 112
                                created: "07/01/2026, 07:53:47.714"
                                modified: "07/01/2026, 07:55:20.829"
                                cartid: 4
                                productid: 4
                                qty: 5
                '400':
                    description: Invalid request data supplied.
                '404':
                    description: Cart item not found.
        delete:
            summary: Delete cart item
            description: Deletes a select item by its unique item ID
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  required: true
                  description: The ID of the user whose cart will be altered.
                  schema:
                    type: string
                - name: cartItemId
                  in: path
                  required: true
                  description: The item ID to be removed from user's cart.
                  schema:
                    type: integer
            responses:
                '204':
                    description: Item successfully deleted.
                '400':
                    description: Bad request. Invalid user ID format.
                '404':
                    description: Cart or Item not found.
    /carts/user/{userId}/checkout:
        summary: User checkout
        description: User pays and creates an order.
        post:
            summary: Checkout user cart
            description: Allow user to checkout
            tags:
                - Carts
            parameters:
                - name: userId
                  in: path
                  description: The ID of the user to add items to cart
                  required: true
                  schema:
                    type: string
            responses:
                '200':
                    description: Order successfully created.
                    content:
                        application/json:
                            example:
                                id: 77
                                created: "2026-07-01T11:58:33.150Z"
                                modified: "2026-07-01T11:58:35.154Z"
                                total: "699.99"
                                status: "COMPLETE"
                                userid: 1
                '400':
                    description: Cannot proceed to checkout. Cart is empty.
                '404':
                    description: Not Found. User or cart could not be found.
                '500':
                    description: Internal Server Error. Database or payment failure.
                    